+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to the "amiga-debug" extension will be documented in this file.
 
+## 1.4.2
+- fixes missing shrinkler terminal output
+- Shrinkler: fixed progress output (only went to 10% instead of 100%)
+
+## 1.4.1
+- fixed DMA display for blitter (Channel D was missing, Channel B and C incorrectly had Line & Fill attribution) (fixes [#104](https://github.com/BartmanAbyss/vscode-amiga-debug/issues/104))
+- added more DMA events (CPU Blitter Steal, CPU Blitter Stolen, Copper Skip)
+- fix Makefile for users with `git sh` in their path (fixes [#103](https://github.com/BartmanAbyss/vscode-amiga-debug/issues/103))
+
+## 1.4.0
+- symbols for Kickstart 1.2, 1.3, 2.04, 2.05, 3.0, 3.1 (no FPU support) working for debugging (e.g. you can set a breakpoint to 'WaitBlit' or 'AllocMem') and profiling
+
+## 1.3.9
+- fix debugger not working when kickstart symbols not found (broken since 1.3.8)
+
+## 1.3.8
+- symbols for Kickstart 1.3 [exec 34.2] (GDB only, dos.library missing)
+
+## 1.3.7
+- fixed crash in assembly view #97
+
+## 1.3.6
+- use VSCode built-in memory viewer for `Amiga: View Memory` 
+
 ## 1.3.5
 - use VSCode built-in memory viewer (click the binary icon beside a variable in the debug view)
 - fix crash in WinUAE when requesting a memory dump